rebase and fixes
